move dsp files to /dsp
