+++ Makefile
@@ -2,7 +2,7 @@
 RACK_DIR ?= ../..
 
 # FLAGS will be passed to both the C and C++ compiler
-FLAGS += -I./sqsrc/util -I./composites -I./src
+FLAGS += -I./sqsrc/util -I./composites -I./src -I./dsp
 CFLAGS +=
 CXXFLAGS +=
